Move node holder to DI
